test: refactor test suite and standardize output naming
- Split monolithic test file into separate test files for each scenario (basic, versioning, sse-s3, sse-kms, folders)
- Rename example directory from `with-folders` to `folders` for consistency
- Update all example outputs to use `versioning_status` instead of `versioning_enabled`
- Update CI workflow to specify test file paths explicitly for improved test isolation
- Add explicit test file references in go test commands to prevent test discovery issues
- Add extra blank line in modules/bucket/main.tf for formatting consistency
- Improves test maintainability and aligns output naming conventions across all examples
