+++ examples/folders/main.tf
