+++ examples/folders/outputs.tf
