+++ examples/folders/variables.tf
