+++ modules/bucket/main.tf
@@ -1,6 +1,7 @@
 # AWS S3 Bucket Resource
 # Creates and manages an S3 bucket with optional encryption, versioning, and folder structure
 
+
 # Look up existing KMS key by alias if provided
 data "aws_kms_alias" "this" {
   count = var.s3_config.kms_key_alias != null ? 1 : 0
